#8 강의소개 for User
