+++ src/App.vue
@@ -1,15 +1,10 @@
 <template>
   <div id="app">
-    <Info></Info>
+
   </div>
 </template>
 <script>
-import Info from './sample/lecture_type/info.vue'
-export default {
-  components : {
-    'Info' : Info
-  }
-}
+
 </script>
 
 
@@ -17,7 +12,6 @@ export default {
 #app {
   margin-top: 1.5rem;
   line-height: 1.4285;
-  width : 60%;
 }
 
 </style>
